Remove CSAPP error-checking wrapper functions
For the particular purpose of tshlab and proxylab, I think this change
is in the students' best interest. The wrapper functions are often not
well understood by students, and the capitalization of the first letter
does not call out to them: DANGER! THIS WILL EXIT YOUR PROCESS! which it
absolutely should.

The use of wrapper functions in a textbook is understandable to avoid
clutter. But personally I think students should learn to error-check C
functions explicitly. This gives them the opportunity to actually
consciously think about what they are doing. Error-checking is painful,
but such is the nature of writing C code.